requirements.txt
# Для работы django-приложений
Django>=3.2.12
pip>=21.1.2
environs>=9.5.0
django-phonenumber-field[phonenumbers]>=7.0.0
django-cors-headers>=3.13.0
djangorestframework>=3.14.0

# Для работы c БД postgres
psycopg2>=2.9.5

#Для отправки запросов на api
requests>=2.28.1

# Для работы телеграмм бота
aiogram>=2.22.2
asyncpg>=0.27.0
aiohttp>=3.8.3